* cli-enhancements:
  Issue pull request for branch cli-enhancements
  Use right grep to find overfull hboxes
  Add the --once option to generate PDFs only once
  Improve template preface
  Add more random template polish
  Fix EPUB validation by updating 2nd & 3rd chapters
  Polish
  Add styles in support of Markdown features
  Finish first cut at a decent Markdown template

+++ lib/polytexnic/builders/html.rb
@@ -19,7 +19,9 @@ module Polytexnic
         end
 
         if manifest.markdown?
-          FileUtils.rm(Dir.glob(path('chapters/*.tex')))
+          unless options[:'find-overfull']
+            FileUtils.rm(Dir.glob(path('chapters/*.tex')))
+          end
           manifest.chapters.each do |chapter|
             write_latex_files(chapter)
           end
